Add support for coroutines
Using the `await` special, a function becomes a coroutine suitable for
use with Python's `asyncio` library.

CPython actually produces a `YIELD_FROM` opcode for `await`, which is
mysteriously always (¿?) preceded by pushing a `None` value that is sent
to the generator, coroutine, or iterator returned by `GET_AWAITABLE` or
`GET_YIELD_FROM_ITER`[0]. With this patch, both the new `await` special
and the `yield-from` special get a `send:` keyword argument that allows
sending values other than `None` with `YIELD_FROM`. This doesn't appear
to be defined behavior, so expect surprises.

[0] https://github.com/python/cpython/blob/b54a99d6432de93de85be2b42a63774f8b4581a0/Python/ceval.c#L2157